API

NestedGraphs.domain — Method

domain(ng::NestedGraph, v::Integer) -> Integer

Get the domain index of a vertex v

NestedGraphs.domainedge — Method

domainedge(
    ng::NestedGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> Union{Nothing, Graphs.SimpleGraphs.SimpleEdge}

Get the domain of an edge If the edge connects 2 domains, it returns nothing

NestedGraphs.edge — Method

edge(
    ng::NestedGraph,
    ce::NestedEdge
) -> Graphs.SimpleGraphs.SimpleEdge{Int64}

Convert a local view NestedEdge to a global view

NestedGraphs.interdomainedges — Method

interdomainedges(ng::NestedGraph) -> Any

Get all edges that have no domain, i.e. that interconnect domains

NestedGraphs.nestededge — Method

nestededge(
    ng::NestedGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> NestedEdge

Convert a global view of an edge to local view NestedEdge

NestedGraphs.nestedvertex — Method

nestedvertex(ng::NestedGraph, v) -> Any

NestedGraphs.roll_vertex — Method

roll_vertex(
    ng::NestedGraph,
    v::AbstractArray{T<:Integer, 1}
) -> Union{Nothing, Int64}

Given a vector of the nested inner domains get the index in the flat graph. The last element of the vector is handled as the node number in the v[1:end-1] inner nested graph

NestedGraphs.unroll_vertex — Method

Unroll a nested vertex along all the nested graph domains

NestedGraphs.vertex — Method

vertex(
    ng::NestedGraph,
    cv::Tuple{T, T} where T<:Integer
) -> Union{Nothing, Int64}

Convert a local view NestedVertex to a global view

NestedGraphs.vertex — Method

vertex(
    ng::NestedGraph,
    d::Integer,
    v::Integer
) -> Union{Nothing, Int64}

NestedGraphs.NestedEdge — Type

struct NestedEdge{T<:Integer} <: Graphs.SimpleGraphs.AbstractSimpleEdge{T<:Integer}

src::Tuple{T, T} where T<:Integer

dst::Tuple{T, T} where T<:Integer

A NestedEdge connects graphs inside a NestedGraph or simply nodes inside a NestedGraph. The NestedEdge connects two NestedVertexs. This means that every NestedEdge connects to a specific node and not as a hyperedge to the whole domain graph.

NestedGraphs.NestedGraph — Type

struct NestedGraph{T<:Integer, R<:AbstractGraph{T<:Integer}, N<:AbstractGraph} <: AbstractGraph{T<:Integer}

flatgr::AbstractGraph{T} where T<:Integer

: Flat graph view combining all domain graphs given with the edges

grv::Vector{N} where N<:AbstractGraph

: Original domain graphs

neds::Array{NestedEdge{T}, 1} where T<:Integer

: interdomain edges

vmap::Array{Tuple{T, T}, 1} where T<:Integer

: Maps the nodes of flat network to the original graph (Domain, Node)

A NestedGraph is a graph of vertices, where each vertex can be a complete graph. Connections are done with NestedEdges and the vertices are NestedVertexs. NestedGraphs of NestedGraphs are allowed.

NestedGraphs.NestedVertex — Type

struct Tuple{T<:Integer, T<:Integer}

A NestedVertex is the local view of a vertex inside a NestedGraph. It contains the domain and the vertex indices. Basically, it's an alias for a 2-element Tuple.

Index

NestedGraphs.NestedEdge

NestedGraphs.NestedGraph

NestedGraphs.NestedVertex

NestedGraphs.domain

NestedGraphs.domainedge

NestedGraphs.edge

NestedGraphs.interdomainedges

NestedGraphs.nestededge

NestedGraphs.nestedvertex

NestedGraphs.roll_vertex

NestedGraphs.unroll_vertex

NestedGraphs.vertex

NestedGraphs.vertex