API

NestedGraphs.edge — Method

edge(
    ng::NestedGraph,
    ce::NestedEdge
) -> Graphs.SimpleGraphs.SimpleEdge{Int64}

Convert a local view NestedEdge to a global view

NestedGraphs.intersubgraphedges — Method

intersubgraphedges(ng::NestedGraph) -> Any

Get all edges that have no subgraph, i.e. that interconnect subgraphs

NestedGraphs.nestededge — Method

nestededge(
    ng::NestedGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> NestedEdge

Convert a global view of an edge to local view NestedEdge

NestedGraphs.nestedvertex — Method

nestedvertex(ng::NestedGraph, v) -> Any

NestedGraphs.roll_vertex — Method

roll_vertex(
    ng::NestedGraph,
    v::AbstractArray{T<:Integer, 1}
) -> Union{Nothing, Int64}

Given a vector of the nested inner subgraphs get the index in the flat graph. The last element of the vector is handled as the node number in the v[1:end-1] inner nested graph

NestedGraphs.subgraph — Method

subgraph(ng::NestedGraph, v::Integer) -> Integer

Get the subgraph index of a vertex v

NestedGraphs.subgraphedge — Method

subgraphedge(
    ng::NestedGraph,
    e::Graphs.SimpleGraphs.SimpleEdge
) -> Union{Nothing, Graphs.SimpleGraphs.SimpleEdge}

Get the subgraph of an edge If the edge connects 2 subgraphs, it returns nothing

NestedGraphs.unroll_vertex — Method

unroll_vertex(ng::NestedGraph, v::Integer) -> Vector

Unroll a nested vertex along all the nested graph subgraphs

NestedGraphs.vertex — Method

vertex(
    ng::NestedGraph,
    cv::Tuple{T, T} where T<:Integer
) -> Union{Nothing, Int64}

Convert a local view NestedVertex to a global view

NestedGraphs.vertex — Method

vertex(
    ng::NestedGraph,
    d::Integer,
    v::Integer
) -> Union{Nothing, Int64}

NestedGraphs.NestedEdge — Type

struct NestedEdge{T<:Integer} <: Graphs.SimpleGraphs.AbstractSimpleEdge{T<:Integer}

src::Tuple{T, T} where T<:Integer

dst::Tuple{T, T} where T<:Integer

A NestedEdge connects graphs inside a NestedGraph or simply nodes inside a NestedGraph. The NestedEdge connects two NestedVertexs. This means that every NestedEdge connects to a specific node and not as a hyperedge to the whole subgraph graph.

NestedGraphs.NestedGraph — Type

struct NestedGraph{T<:Integer, R<:AbstractGraph{T<:Integer}, N<:AbstractGraph} <: AbstractGraph{T<:Integer}

flatgr::AbstractGraph{T} where T<:Integer: Flat graph view combining all subgraph graphs given with the edges

grv::Vector{N} where N<:AbstractGraph: Original subgraph graphs

neds::Array{NestedEdge{T}, 1} where T<:Integer: intersubgraph edges

vmap::Array{Tuple{T, T}, 1} where T<:Integer: Maps the nodes of flat network to the original graph (Subgraph, Node)

A NestedGraph is a graph of vertices, where each vertex can be a complete graph. Connections are done with NestedEdges and the vertices are NestedVertexs. NestedGraphs of NestedGraphs are allowed.

NestedGraphs.NestedGraph — Method

NestedGraph(
    grv::Array{R<:AbstractGraph, 1},
    edges::AbstractVector;
    both_ways
) -> NestedGraph{Int64, _A} where _A<:AbstractGraph{Int64}

both_ways controls whether edges should be added also in reverse.

NestedGraphs.NestedGraph — Method

extrasubgraph controls the NestedGraph should be initialized with an empty subgraph

NestedGraphs.NestedVertex — Type

struct Tuple{T<:Integer, T<:Integer}

A NestedVertex is the local view of a vertex inside a NestedGraph. It contains the subgraph and the vertex indices. Basically, it's an alias for a 2-element Tuple.

Index

NestedGraphs.NestedEdge

NestedGraphs.NestedGraph

NestedGraphs.NestedGraph

NestedGraphs.NestedGraph

NestedGraphs.NestedVertex

NestedGraphs.edge

NestedGraphs.intersubgraphedges

NestedGraphs.nestededge

NestedGraphs.nestedvertex

NestedGraphs.roll_vertex

NestedGraphs.subgraph

NestedGraphs.subgraphedge

NestedGraphs.unroll_vertex

NestedGraphs.vertex

NestedGraphs.vertex